+++ Changelog.md
@@ -2,6 +2,7 @@
 <summary> 1.0.1 </summary>
 
 - Fixed issue with ContainerBehaviour scripts not resetting the movement debuffs when the player dies holding one.
+- Fixed issue with Lategame Upgrades Compatibility not being correctly implemented.
 
 
 </details>